projects/src/tokscraper/pyproject.toml
[project]
name = "tokscraper"
version = "0.1.0"
description = "Scrapes Tiktok, unsurprisingly"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bgutil-ytdlp-pot-provider>=1.2.2",
    "curl-cffi>=0.13.0",
    "google-cloud-storage>=3.1.1",
    "pas-log",
    "pydantic>=2.11.7",
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "structlog>=25.4.0",
    "yt-dlp>=2025.08.27",
]

[tool.uv]
package = true

[tool.uv.sources]
pas-log = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = ["responses>=0.25.7"]

[tool.hatch.build.targets.wheel]
packages = ["tokscraper"]
